Consolidate getOutputDisplay() within nested-workflow-base
